data: rename four sample EDIDs
The Dell EDIDs didn't conform to the EDID naming in this directory.
The same was true for the LG EDID. I'm guessing it was tested with
the DVI connector, not the VGA connector. Unfortunately, this was
never specified in the original commit, so this is just a guess.

The elo-4600L-hdmi EDID contained an upper case L which should be
a lower case l.

Also documented the naming convention in the README file.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>