edid-decode: round down frame and horizontal rates
When checking framerates and horizontal frequencies against the
specified monitor range assume that these values should be rounded
down.

E.g. if the refresh rate is 60.317 and the max monitor rate is 60,
then round down the refresh rate to 60.

The EDID spec explicitly mentions that the max monitor pixelclock
rate should be rounded *up*, which implies that the other ranges
are actually rounded down from the maximum supported values.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>