edid-decode: check for presence of VCDB
The Video Capability Data Block is needed to report Selectable RGB
Quantization Range. A warning was present to report if Selectable
RGB Quantization Range was not supported in the VCDB since that gives
interop problems. But that warning did not trigger if the VCDB was
completely missing.

Add a check for that and issue a warning if it is missing.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>