edid-decode: cta.preferred_timings wasn't cleared
If the first svd can be the preferred timing, then clear
cta.preferred_timings before adding that SVD. It overrides
the regular DTD, so any existing preferred timings must be
cleared first.

Also set first_svd_might_be_preferred to false once this is
done to avoid this to be repeated if there are  SVDs in multiple
CTA extension blocks.

And never set first_svd_might_be_preferred if there is a VFPDB.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>