edid-decode: improve error text and add comments
Improve error messages that refer to Byte 3 of the CTA Extension
some more, and add a comment to Table 52 of CTA-861-G that describes
this byte.

It's annoying that these bits in that byte never got a symbolic name
in the standard, so it is hard to find a good way to refer to them
in the error message.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>