edid-decode: improve pref timing warning, downgrade fail() to warn()
The 'VIC 97 is the preferred timing, overriding the first detailed
timings. Is this intended?' message is confusing if the VIC and the
first DTD have matching timings. In that case change the message to
'For improved preferred timing interoperability, set 'Native detailed
modes' to 1.'.

Also downgrade the 'Set Selectable YCbCr Quantization to avoid interop
issues.' message to a warn(). Too many EDIDs do not set this, and the
interop risks are a lot lower compared to RGB Quantization.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>